+++ .circleci/config.yml
@@ -115,7 +115,6 @@ jobs:
       # NOTE: because this doesn't build with the kubernetes tag, it will not run the kubernetes tests. See
       # kubernetes_test build steps.
       - run: mkdir -p /tmp/logs
-      - run: run-go-tests --packages "-p 1 -tags=special -run TestRunCommandWithHugeLineOutput ./modules/shell"
       # check we can compile the azure code, but don't actually run the tests
       - run: run-go-tests --packages "-p 1 -tags=azure -run IDontExist ./modules/azure"
       - run: run-go-tests --packages "-p 1 ./..." | tee /tmp/logs/test_output.log
